requirements.txt
autopep8==2.0.1
camelcase==0.2
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
pycodestyle==2.10.0
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.46
Werkzeug==2.2.2
